// src/app/services/models/weather-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface WeatherDto {
  chanceOfRain?: number;
  chanceOfSnow?: number;
  cityName?: string;
  country?: string;
  date?: string;
  maxTemperature?: number;
  maxWind?: number;
  minTemperature?: number;
  sunrise?: string;
  sunset?: string;
}